backend/requirements.txt
uvicorn
python-multipart
google-generativeai
fastapi
openpyxl
python-dotenv

